Update to v0.10 (#16)
* GERTi v0.9.1

Changelog:
1. Replace previous system of attempting to use a single connection for
bi-directional transfer. Now automatically adds and utilizes two
uni-directional connections
2. Replace "nil" "nillerino" system with a more compact system that
better suits its purpose
3. Inserted blank lines to make code more readable
4. Significantly reduced the number of "if...then" statements in GERTi
Gateway

* Assign GERTe number to viomi

* Update GERTi Client with modem checking (#8)

* Update GERTi Client.lua

Error out of the program if no tier of network card is found.

*Author's note - This -should- work, but testing should be done before applying to main branch.

* Tweak changes slightly

* Combine modem initialization code

* GERTi v0.9.2

Big changes. Code neatening, performance increases, and more! Check the
changelog for a full list.

* GERTi v0.9.2B

Brings the v0.9.2 changes to the Gateway, and a quick bugfix with the
handler code for the client.

* Card Checking + Bug Fixes (#9)

* Check if modem and/or tunnel, check isWireless(), handler fix, couple of typoes fixed, minor bug fixes, commented out undefined handleEvent()

* Better tunnel vs. modem checks

* Commented out another undefined handleEvent(), another error message for modem vs. tunnel usage.

* Requested Changes Made

Changes previously made to Client now synchronized to Gateway

require tunnel and modem

* Fix Requires (#10)

* Fix requires

* Fix Requires

* Add 2 reservations by nxsupert

* Changed README.md (#11)

OC-Programs is an entirely different repo.

* GERTi v0.9.3

This update fixes a bunch of bugs, but a critical data transmission bug
is not fixed

* Enhancements and Re-Organizations

* Changed README.md

OC-Programs is an entirely different repo.

* DATA return true

* Spaces < Tabs

* OPENROUTE Handler Changes

When a route is found/opened, return true so you don't test a bunch of if statements we already know are false.

Some other small changes.

* Return function calls

* Added similar changes from previous commit to Gateway

* OPENROUTE Rewrite

openRoute() created so we're not rewriting the same code out over and over when it could be functionalized.

Got rid of tons of if statements to see if the last for loop found something, replaced w/ just running the code as soon as it's found and returning something.

* DATA Changes Gateway > Client

Brought DATA changes made to Gateway in last commit over to Client

* Bugfix

Forgot to change passed 'port' in openRoute() to 'storePort' so as not to be overwritten on line 129.

Change has already been applied to the Client.

* Removed key1, key2

Already removed from Gateway, forgot to do so with Client.

* Reserve Port 4096 for Izaya

MultICE networking

* v0.10.0-RC1

About freaking time

* Update Whitepaper for GERTe

* Update Whitepaper

Update whitepaper to be *closer* to what GERT actually is with GERTe changes. Also removed advertising, they're already here.

* Update GERTe Whitepaper for GEDS

* Phase 1 Update of Whitepaper

This updates the whitepaper to bring it more in line with how things currently operate, and to communicate clearer. Phase 2 due upon GERT v1.0 completion.

* Minor changes

Reworded and re-arranged the introduction a little bit to make it less awkward.

* Reword Compatibility section

* Rework introduction
