+++ GERTi/Changelog.txt
@@ -1,0 +1,17 @@
+v0.10
+Changed: Instead of calling socket.read/write, socket:read()/write must be used, but they do not need to be provided with information now
+
+Lots of bugfixes
+
+
+v0.9.2
+Changed:
+1. Replaced previous system of attempting to use a single connection for bi-directional data transfer. Now GERTi automatically adds and utilizes two uni-directional connections.
+2. Replace "nil" "nillerino" system with the new handler system
+3. Replaced long if...elseif chain in receive packet in Client/Gateway with new handler system
+4. Replaced modem/tunnel checking system with a new one (Credit to Kodos for that), and centralized modem setup code.
+
+Added:
+1. Added strategic whitespace to make the code more legible
+2. Improved quality and quantity of documentation
+3. New packet handler table system to improve speed of client and gateway, and reduce code complexity
