+++ README.md
@@ -1,4 +1,4 @@
-# OC-Programs
+# Global Empire Routing Technology
 Home of the GERT networking technology. More detailed information can be found in the GERT whitepaper housed in this repository.
 
 Please read the license
